+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 + [x] Improved rule `date` for validator
 + [x] Added component `resizable-container`
 + [x] Fixed bug in `hotkey` component when using `$(...).hotkey()` method for non-input elements
++ [x] Fixed clear button action in the `select` component